Group B: rename manager files to canonical names with __getattr__ shims
The architecture proposal called for the canonical Layer-3 manager
classes to live at module names that match the class name
(``node_manager.py`` etc.), with the legacy module names becoming
thin compat shims that emit ``DeprecationWarning`` when the legacy
class name is imported. PR #46 established that pattern for the two
pure compat shim files (Group A); this PR completes it for the four
manager modules (Group B).

File renames
------------
- nodes/nodes.py        -> nodes/node_manager.py
- elements/elements.py  -> elements/element_manager.py
- model/model_info.py   -> model/model_info_reader.py
- model/cdata.py        -> model/cdata_reader.py

The ``OldName = NewName`` alias is removed from each canonical file
and reinstated on the package __init__ (quiet) and on a new shim file
at the legacy module path (warns via PEP 562 ``__getattr__``).

Behavior
--------
| Import                                                | After PR        |
|-------------------------------------------------------|-----------------|
| from STKO_to_python import Nodes                      | quiet           |
| from STKO_to_python.nodes import NodeManager, Nodes   | quiet           |
| from STKO_to_python.nodes.node_manager import ...     | quiet           |
| from STKO_to_python.nodes.nodes import NodeManager    | quiet (re-export)|
| from STKO_to_python.nodes.nodes import Nodes          | DeprecationWarning |

The legacy class name resolves to the same canonical class object in
every case — pickle compat and ``isinstance`` checks are preserved.

Internal updates
----------------
- src/STKO_to_python/__init__.py and core/dataset.py import canonical
  names directly so the library never trips its own deprecation.
- One internal self-reference (``Nodes._sort_step_keys`` inside the
  static method on what is now NodeManager) is corrected to use the
  canonical class name.

Tests
-----
- test_manager_aliases.py rewritten: each legacy-name import is now
  wrapped in pytest.warns(DeprecationWarning); canonical imports run
  under a strict ``error::DeprecationWarning`` filter to prove they
  stay quiet.
- test_public_api.py: legacy deep paths moved into the
  DEPRECATED_DEEP_IMPORTS parametrize block (4 new entries); the
  canonical post-Group-B paths are added to PUBLIC_SYMBOLS.
- test_logging.py: LOGGING_MODULES updated to point at the new
  canonical module names where the loggers actually live.
- test_nodes_flatten.py and test_element_bucket_grouping.py switch
  their deep imports to the canonical paths.

Full suite: 628 passed.
