+++ src/STKO_to_python/elements/__init__.py
@@ -1,5 +1,10 @@
-from .elements import ElementManager, Elements
+from .element_manager import ElementManager
 from .element_results import ElementResults
+
+# Back-compat alias preserved on the package surface (quiet); the deep
+# path ``STKO_to_python.elements.elements.Elements`` emits a
+# ``DeprecationWarning`` via PEP 562 ``__getattr__``.
+Elements = ElementManager
 
 __all__ = [
     'ElementManager',
