+++ src/STKO_to_python/model/__init__.py
@@ -1,5 +1,12 @@
-from .model_info import ModelInfo, ModelInfoReader
-from .cdata import CData, CDataReader
+from .model_info_reader import ModelInfoReader
+from .cdata_reader import CDataReader
+
+# Back-compat aliases preserved on the package surface (quiet); the
+# deep paths ``STKO_to_python.model.model_info.ModelInfo`` and
+# ``STKO_to_python.model.cdata.CData`` emit ``DeprecationWarning`` via
+# PEP 562 ``__getattr__``.
+ModelInfo = ModelInfoReader
+CData = CDataReader
 
 __all__ = [
     "ModelInfo",
