+++ src/STKO_to_python/model/model_info_reader.py
@@ -1,0 +1,752 @@
+import glob
+import os
+import re
+from typing import TYPE_CHECKING
+from collections import defaultdict
+from typing import Optional, Dict, List, Sequence, Any
+import numpy as np
+import pandas as pd
+import logging
+
+logger = logging.getLogger(__name__)
+
+if TYPE_CHECKING:
+    from ..core.dataset import MPCODataSet
+
+class ModelInfoReader:
+    """
+    Canonical Layer 3 reader for MPCO model metadata.
+
+    Loads once on construction (via the ``MPCODataSet`` friend
+    relationship), then exposes read-only views. The legacy name
+    ``ModelInfo`` is preserved as an alias at the bottom of this module.
+    New code should prefer ``ModelInfoReader``.
+
+    This class has a "friend" relationship with MPCODataSet, which is allowed
+    to access protected methods.
+    """
+    def __init__(self, dataset:'MPCODataSet'):
+        self.dataset = dataset
+        # Lightweight collaborator — stateless, no construction cost.
+        from ..io.time_series_reader import TimeSeriesReader
+        self._time_series_reader = TimeSeriesReader()
+
+    def _get_file_list(self, extension: Optional[str] = None, verbose: bool = False) -> Dict[str, Dict[int, str]]:
+        """
+        Retrieves a mapping of partitioned files from the dataset directory.
+
+        Parameters
+        ----------
+        extension : str, optional
+            The file extension to look for (e.g., 'mpco', 'cdata'). If not provided,
+            it uses the extension from `self.MCPODataSet.file_extension`.
+        verbose : bool, optional
+            If True, prints detailed information about the files found.
+
+        Returns
+        -------
+        dict
+            A nested dictionary structured as:
+            {
+                'recorder_base_name': {
+                    0: '/path/to/recorder_base_name.part-0.<ext>',
+                    1: '/path/to/recorder_base_name.part-1.<ext>',
+                    ...
+                },
+                ...
+            }
+
+        Raises
+        ------
+        FileNotFoundError
+            If no files with the given extension are found in the specified directory.
+        Exception
+            For unexpected errors during file parsing.
+        """
+        if extension is None:
+            extension = self.dataset.file_extension.strip("*.")
+
+        results_directory = self.dataset.hdf5_directory
+
+        try:
+            files = glob.glob(os.path.join(results_directory, f"*.{extension}"))
+            if not files:
+                raise FileNotFoundError(f"No .{extension} files found in {results_directory}")
+
+            file_mapping = defaultdict(dict)
+
+            for file in files:
+                filename = os.path.basename(file)
+
+                if ".part-" in filename:
+                    try:
+                        name, part_str = filename.split(".part-", 1)
+                        part = int(part_str.split(".")[0])
+                        file_mapping[name][part] = file
+                    except (ValueError, IndexError):
+                        logger.warning("Skipping file due to unexpected naming format: %s", file)
+                else:
+                    # Handle compound extensions like ".mpco.cdata"
+                    if filename.endswith(f".{extension}"):
+                        # Remove .<extension> (e.g., .cdata)
+                        base_with_possible_extra_ext = filename[: -len(extension) - 1]
+                        # Remove any remaining extra extension like .mpco if present
+                        base = re.sub(r"\.mpco$", "", base_with_possible_extra_ext)
+                        file_mapping[base][0] = file
+
+            if verbose:
+                logger.info("Found files:")
+                for name, parts in file_mapping.items():
+                    logger.info("%s:", name)
+                    for part, path in sorted(parts.items()):
+                        logger.info("  Part: %s, File: %s", part, path)
+
+            return file_mapping
+
+        except Exception as e:
+            logger.error("Model Info Error during file listing: %s", e)
+            raise
+        
+    def _get_file_list_for_results_name(self, extension= None, verbose=False):
+
+        if extension is None:
+            extension=self.dataset.file_extension.strip("*.")
+            
+        file_info=self._get_file_list(extension=extension, verbose=verbose)
+
+        recorder_files = file_info.get(self.dataset.recorder_name)
+
+        if recorder_files is None:
+            raise ValueError(f"Model Info Error: Recorder name '{self.dataset.recorder_name}' not found in {extension} files.")
+
+        return recorder_files
+    
+    def _get_model_stages(self, verbose=False):
+        """
+        Retrieve model stages from all result partitions.
+
+        Args:
+            verbose (bool, optional): If True, prints the model stages.
+
+        Returns:
+            list: Sorted list of model stage names from all partitions.
+        """
+        model_stages = []
+        policy = self.dataset._format_policy
+
+        # Use partition paths from the dictionary created by _get_results_partitions
+        for part_idx in self.dataset.results_partitions:
+            with self.dataset._pool.with_partition(part_idx) as results:
+                # Get model stages from the current partition file
+                partition_stages = [
+                    key for key in results.keys()
+                    if policy.is_model_stage_group(key)
+                ]
+                model_stages.extend(partition_stages)
+
+        # Remove duplicates by converting to a set, then back to a sorted list
+        model_stages = sorted(set(model_stages))
+
+        if not model_stages:
+            raise ValueError("Model Info Error: No model stages found in the result partitions.")
+
+        if verbose:
+            logger.info('The model stages found across partitions are: %s', model_stages)
+
+        return model_stages
+    
+    def _get_node_results_names(
+            self,
+            model_stage: Optional[str] = None,
+            verbose: bool = False,
+            raise_if_empty: bool = False
+    ) -> List[str]:
+        """
+        Retrieve the names of nodal results for a given model stage.
+
+        Args:
+            model_stage (str, optional): Model stage name. If None, search all stages.
+            verbose (bool, optional): Print the discovered result names.
+            raise_if_empty (bool, optional): If True, raise ValueError when nothing is found.
+                                            If False (default) return an empty list instead.
+
+        Returns:
+            list[str]: Sorted list of nodal result names (may be empty).
+        """
+        # 1. Determine which stages to inspect
+        model_stages = [model_stage] if model_stage else self.dataset.model_stages
+
+        node_results_names: set[str] = set()
+        policy = self.dataset._format_policy
+
+        # 2. Scan every partition for every requested stage
+        for stage in model_stages:
+            for part_idx in self.dataset.results_partitions:
+                with self.dataset._pool.with_partition(part_idx) as results:
+                    nodes_group = results.get(policy.results_on_nodes_path(stage))
+                    if nodes_group:
+                        node_results_names.update(nodes_group.keys())
+
+            if verbose:
+                logger.info("Node results in '%s': %s", stage, sorted(node_results_names))
+
+        # 3. Handle empty results according to caller’s wishes
+        if not node_results_names:
+            message = (
+                f"No nodal results found for stage(s): {', '.join(model_stages)} "
+                f"in {len(self.dataset.results_partitions)} partition(s)."
+            )
+            if raise_if_empty:
+                raise ValueError(message)
+            logger.warning(message)
+
+        return sorted(node_results_names)
+    
+    def _get_elements_results_names(
+            self,
+            model_stage: Optional[str] = None,
+            verbose: bool = False,
+            raise_if_empty: bool = False
+    ) -> list[str]:
+        """
+        Retrieve the names of element results for a given model stage.
+
+        Args:
+            model_stage (str, optional): Name of the model stage. If None, search all stages.
+            verbose (bool, optional): Print the discovered result names.
+            raise_if_empty (bool, optional): If True, raise ValueError when nothing is found.
+                                            If False (default) return an empty list instead.
+
+        Returns:
+            list[str]: Sorted list of element result names (may be empty).
+        """
+        # 1. Determine which stages to inspect
+        model_stages = [model_stage] if model_stage else self.dataset.model_stages
+
+        element_results_names: set[str] = set()
+        policy = self.dataset._format_policy
+
+        # 2. Scan every partition for every requested stage
+        for stage in model_stages:
+            for part_idx in self.dataset.results_partitions:
+                with self.dataset._pool.with_partition(part_idx) as results:
+                    ele_group = results.get(policy.results_on_elements_path(stage))
+                    if ele_group:
+                        element_results_names.update(ele_group.keys())
+
+            if verbose:
+                logger.info("Element results in '%s': %s", stage, sorted(element_results_names))
+
+        # 3. Handle empty results according to caller’s wishes
+        if not element_results_names:
+            message = (
+                f"No element results found for stage(s): {', '.join(model_stages)} "
+                f"in {len(self.dataset.results_partitions)} partition(s)."
+            )
+            if raise_if_empty:
+                raise ValueError(message)
+            logger.warning(message)
+
+        return sorted(element_results_names)
+
+    def _get_element_types(
+            self,
+            model_stage: Optional[str] = None,
+            results_name: Optional[str] = None,
+            *,
+            verbose: bool = False,
+            raise_if_empty: bool = False,
+    ) -> dict[str, Any]:
+        """
+        Retrieve element types (full decorated names) per element-results group, across
+        the requested model stages and all partitions.
+
+        Behavior:
+        - If `model_stage` is None, inspects all stages in `self.dataset.model_stages`.
+        - If `results_name` is None, discovers element-results names per stage using
+        `_get_elements_results_names(stage)`.
+        - Skips stages with no element results.
+        - Aggregates across partitions; deduplicates; returns sorted outputs.
+        - Does not raise when nothing is found unless `raise_if_empty=True`.
+
+        Args:
+            model_stage (str, optional): Stage to query; if None, query all stages.
+            results_name (str, optional): Specific element-results group to query; if None,
+                                        all groups found per stage are used.
+            verbose (bool): Print progress/details.
+            raise_if_empty (bool): Raise ValueError if nothing found.
+
+        Returns:
+            dict[str, Any]:
+                {
+                    "element_types_dict": dict[str, list[str]],  # result_name -> sorted list of types
+                    "unique_element_types": list[str],           # sorted unique across all results
+                    "skipped_stages": list[str],                 # stages skipped due to emptiness
+                    "scanned_stages": list[str],                 # stages actually scanned
+                }
+        """
+        stages = [model_stage] if model_stage else list(self.dataset.model_stages)
+
+        element_types_dict: dict[str, set[str]] = {}
+        skipped_stages: list[str] = []
+        scanned_stages: list[str] = []
+
+        for stage in stages:
+            # Determine which result groups to look at for this stage
+            if results_name is None:
+                result_names = self._get_elements_results_names(
+                    model_stage=stage, verbose=False, raise_if_empty=False
+                )
+            else:
+                result_names = [results_name]
+
+            if not result_names:
+                skipped_stages.append(stage)
+                if verbose:
+                    logger.info("Skipping stage '%s': no element results found.", stage)
+                continue
+
+            scanned_stages.append(stage)
+            policy = self.dataset._format_policy
+
+            # Walk all partitions, gather element type groups per result name
+            for part_idx in self.dataset.results_partitions:
+                with self.dataset._pool.with_partition(part_idx) as results:
+                    base_group = results.get(policy.results_on_elements_path(stage))
+                    if base_group is None:
+                        # This partition has no elements group for this stage
+                        continue
+
+                    for name in result_names:
+                        grp = base_group.get(name)
+                        if grp is None:
+                            # This partition lacks this particular result group
+                            continue
+
+                        # Initialize set for this result group
+                        if name not in element_types_dict:
+                            element_types_dict[name] = set()
+
+                        # Keep FULL decorated element type names
+                        element_types_dict[name].update(grp.keys())
+
+            if verbose:
+                counts = {k: len(v) for k, v in element_types_dict.items()}
+                logger.info(
+                    "Stage '%s': collected types (counts per result) -> %s",
+                    stage, counts,
+                )
+
+        # Finalize / convert sets -> sorted lists
+        element_types_dict_sorted: dict[str, list[str]] = {
+            name: sorted(types_set) for name, types_set in element_types_dict.items()
+        }
+
+        # Unique across all
+        unique_all: list[str] = sorted(
+            set().union(*element_types_dict_sorted.values()) if element_types_dict_sorted else set()
+        )
+
+        # Handle emptiness policy
+        if not element_types_dict_sorted:
+            msg = (
+                "No element types found. "
+                f"Checked stages: {', '.join(stages)}. "
+                f"Skipped (empty) stages: {', '.join(skipped_stages) if skipped_stages else 'none'}. "
+                f"Partitions scanned: {len(self.dataset.results_partitions)}."
+            )
+            if raise_if_empty:
+                raise ValueError(f"Model Info: {msg}")
+            logger.warning(f"Model Info: {msg}")
+
+        if verbose:
+            logger.info("Unique element types (%d): %s", len(unique_all), unique_all)
+
+        return {
+            "element_types_dict": element_types_dict_sorted,
+            "unique_element_types": unique_all,
+            "skipped_stages": skipped_stages,
+            "scanned_stages": scanned_stages,
+        }
+
+    
+    def _get_all_types(
+            self,
+            model_stage: Optional[str] = None,
+            *,
+            verbose: bool = False,
+            raise_if_empty: bool = False,
+    ) -> list[str]:
+        """
+        Collect ALL element 'type' groups present under every element-result name
+        across the requested model stages and partitions.
+
+        - Skips model stages that have no element results.
+        - Uses full decorated element type names (e.g., '203-ASDShellQ4[201:0:0]').
+
+        Args:
+            model_stage (str, optional): Single stage to query. If None, query all.
+            verbose (bool): If True, prints per-stage info.
+            raise_if_empty (bool): If True, raises if nothing is found.
+
+        Returns:
+            list[str]: Sorted unique element type names (decorated).
+        """
+        # 1) Which stages?
+        stages = [model_stage] if model_stage else self.dataset.model_stages
+
+        element_types: set[str] = set()
+        skipped_stages: list[str] = []
+
+        # 2) For each stage, get its element result names once
+        for stage in stages:
+            result_names = self._get_elements_results_names(
+                model_stage=stage,
+                verbose=False,
+                raise_if_empty=False,  # don't raise here; we'll decide below
+            )
+            if not result_names:
+                skipped_stages.append(stage)
+                if verbose:
+                    logger.info("Skipping stage '%s': no element results found.", stage)
+                continue  # nothing to do for this stage
+
+            # 3) Walk all partitions and harvest element type groups per result name
+            policy = self.dataset._format_policy
+            for part_idx in self.dataset.results_partitions:
+                with self.dataset._pool.with_partition(part_idx) as results:
+                    base_group = results.get(policy.results_on_elements_path(stage))
+                    if base_group is None:
+                        # This partition has no elements group for this stage; skip
+                        continue
+
+                    for name in result_names:
+                        grp = base_group.get(name)
+                        if grp is None:
+                            # This partition lacks this particular result name; skip
+                            continue
+                        # IMPORTANT: keep full decorated type names
+                        element_types.update(grp.keys())
+
+            if verbose:
+                found = sorted(element_types)
+                logger.info("Stage '%s': collected %d type(s) so far.", stage, len(found))
+
+        # 4) Finalize
+        if not element_types:
+            msg = (
+                "No element types found. "
+                f"Checked stages: {', '.join(stages)}. "
+                f"Skipped (empty) stages: {', '.join(skipped_stages) if skipped_stages else 'none'}. "
+                f"Partitions scanned: {len(self.dataset.results_partitions)}."
+            )
+            if raise_if_empty:
+                raise ValueError(f"Model Info: {msg}")
+            logger.warning(f"Model Info: {msg}")
+            return []
+
+        return sorted(element_types)
+
+    
+    def _get_time_series_on_nodes_for_stage(self, model_stage, results_name):
+        """
+        Retrieve and consolidate the unique time series data across all partitions 
+        for a given model stage and nodal results name, returning a Pandas DataFrame.
+
+        Args:
+            model_stage (str): The model stage to query.
+            results_name (str): The nodal results name to query.
+
+        Returns:
+            pd.DataFrame: A DataFrame with columns ['STEP', 'TIME'], sorted by STEP.
+        """
+
+        time_series_dict: dict[int, float] = {}
+
+        for part_number in self.dataset.results_partitions:
+            try:
+                with self.dataset._pool.with_partition(part_number) as partition:
+                    base_path = f"{model_stage}/RESULTS/ON_NODES/{results_name}/DATA"
+                    time_series_dict.update(
+                        self._time_series_reader.read_step_time_pairs(partition.get(base_path))
+                    )
+
+            except Exception as e:
+                logger.error(
+                    "Model Info Error: Get time series error processing partition %s "
+                    "for model stage '%s', results name '%s': %s",
+                    part_number, model_stage, results_name, e,
+                )
+
+        return pd.DataFrame(
+            list(time_series_dict.items()), columns=['STEP', 'TIME']
+        ).sort_values(by='STEP')
+
+    def _get_time_series_on_elements_for_stage(self, model_stage, results_name, element_type):
+        """
+        Retrieve and consolidate the unique time series data across all partitions 
+        for a given model stage, element results name, and specific element type, 
+        returning a Pandas DataFrame.
+
+        Args:
+            model_stage (str): The model stage to query.
+            results_name (str): The element results name to query (e.g., 'force', 'deformation').
+            element_type (str): The specific element type to query (e.g., '203-ASDShellQ4').
+
+        Returns:
+            pd.DataFrame: A DataFrame with columns ['STEP', 'TIME'], sorted by STEP.
+        """
+
+        time_series_dict: dict[int, float] = {}
+
+        for part_number in self.dataset.results_partitions:
+            try:
+                with self.dataset._pool.with_partition(part_number) as partition:
+                    base_path = f"{model_stage}/RESULTS/ON_ELEMENTS/{results_name}/{element_type}/DATA"
+                    time_series_dict.update(
+                        self._time_series_reader.read_step_time_pairs(partition.get(base_path))
+                    )
+
+            except Exception as e:
+                logger.error(
+                    "Model Info Error: Get time series error processing partition %s "
+                    "for model stage '%s', results name '%s', element type '%s': %s",
+                    part_number, model_stage, results_name, element_type, e,
+                )
+
+        # Convert to DataFrame
+        df = pd.DataFrame(list(time_series_dict.items()), columns=['STEP', 'TIME']).sort_values(by='STEP')
+
+        return df
+    
+    def _get_time_series(self) -> pd.DataFrame:
+        """
+        Consolidate the unique STEP–TIME pairs for every model stage,
+        even if a stage contains only nodal *or* only element results.
+
+        Returns
+        -------
+        pd.DataFrame
+            Multi-index ['MODEL_STAGE', 'STEP'] → ['TIME'].
+        """
+
+        # ── convenience ─────────────────────────────────────────────────────────
+        node_names: list[str] = (self.dataset.node_results_names or [])
+        elem_names: list[str] = (self.dataset.element_results_names or [])
+        elem_types_dict: dict[str, list[str]] = self.dataset.element_types.get(
+            "element_types_dict", {}
+        )
+
+        all_time_series: list[pd.DataFrame] = []
+
+        for stage in self.dataset.model_stages:
+            time_df: pd.DataFrame | None = None
+
+            # 1) Try every nodal result (if any) until one yields data
+            for n_result in node_names:
+                df = self._get_time_series_on_nodes_for_stage(stage, n_result)
+                if not df.empty:
+                    time_df = df
+                    break                                       # ← success!
+
+            # 2) If still empty, try the element results
+            if time_df is None or time_df.empty:
+                for e_result in elem_names:
+                    e_types = elem_types_dict.get(e_result, [])
+                    for e_type in e_types:                       # try each element type
+                        df = self._get_time_series_on_elements_for_stage(
+                            stage, e_result, e_type
+                        )
+                        if not df.empty:
+                            time_df = df
+                            break
+                    if time_df is not None and not time_df.empty:
+                        break                                   # ← success!
+            # 3) No data at all → raise a *stage-specific* error
+            if time_df is None or time_df.empty:
+                raise ValueError(
+                    f"Model Info Error: No time-series data found for model stage "
+                    f"'{stage}'. Checked {len(node_names)} nodal results and "
+                    f"{len(elem_names)} element results."
+                )
+
+            # tag with stage and collect
+            time_df["MODEL_STAGE"] = stage
+            all_time_series.append(time_df)
+
+        # ── union ───────────────────────────────────────────────────────────────
+        # ``copy=`` is deprecated in pandas 3.x (copy-on-write is now the
+        # default); passing it emits Pandas4Warning which bubbles up as a
+        # DeprecationWarning under our strict-warning filter.
+        final_df = (
+            pd.concat(all_time_series)
+            .set_index(["MODEL_STAGE", "STEP"])
+            .sort_index()
+        )
+        return final_df
+    
+    def _get_number_of_steps(self) -> Dict[str, int]:
+        """
+        Determine how many analysis steps exist in each model stage,
+        regardless of whether the data are stored under ON_NODES or ON_ELEMENTS.
+
+        Returns
+        -------
+        dict[str, int]
+            Mapping {MODEL_STAGE: n_steps}.
+        """
+        # convenience handles -------------------------------------------------
+        node_names: List[str] = self.dataset.node_results_names or []
+        elem_names: List[str] = self.dataset.element_results_names or []
+        elem_types_dict: Dict[str, List[str]] = self.dataset.element_types.get(
+            "element_types_dict", {}
+        )
+        partition_indices = list(self.dataset.results_partitions)
+
+        steps_info: Dict[str, int] = {}
+
+        for stage in self.dataset.model_stages:
+            step_ids: set[int] = set()
+
+            # 1) nodal results ------------------------------------------------
+            for n_res in node_names:
+                for part_idx in partition_indices:
+                    with self.dataset._pool.with_partition(part_idx) as f:
+                        grp = f.get(f"{stage}/RESULTS/ON_NODES/{n_res}/DATA")
+                        if grp is not None:
+                            step_ids.update(self._to_step_int(k) for k in grp.keys())
+                if step_ids:
+                    break  # found data → stop searching nodal
+
+            # 2) element results ---------------------------------------------
+            if not step_ids:
+                for e_res in elem_names:
+                    for e_type in elem_types_dict.get(e_res, []):
+                        for part_idx in partition_indices:
+                            with self.dataset._pool.with_partition(part_idx) as f:
+                                grp = f.get(
+                                    f"{stage}/RESULTS/ON_ELEMENTS/{e_res}/{e_type}/DATA"
+                                )
+                                if grp is not None:
+                                    step_ids.update(
+                                        self._to_step_int(k) for k in grp.keys()
+                                    )
+                        if step_ids:
+                            break
+                    if step_ids:
+                        break
+
+            # 3) error if nothing found --------------------------------------
+            if not step_ids:
+                raise ValueError(
+                    f"Model Info Error: no STEP datasets located for model stage "
+                    f"'{stage}'. Checked {len(node_names)} nodal results and "
+                    f"{len(elem_names)} element results."
+                )
+
+            steps_info[stage] = len(step_ids)
+
+        return steps_info
+
+    def get_node_coordinates(
+        self,
+        *,
+        node_ids: Sequence[int] | None = None,
+        selection_set_id: int | None = None,
+        as_dict: bool = False,
+    ) -> pd.DataFrame | dict[int, dict[str, Any]]:
+        """
+        Return full rows (node_id, file_id, index, x, y, z, …) from
+        ``self.dataset.nodes_info``.
+
+        Exactly **one** of *node_ids* or *selection_set_id* is required.
+        """
+        # --- XOR check ---------------------------------------------------- #
+        if (node_ids is None) == (selection_set_id is None):
+            raise ValueError(
+                "Specify **either** 'node_ids' **or** 'selection_set_id' (one, not both)."
+            )
+
+        # --- resolve IDs -------------------------------------------------- #
+        if node_ids is None:
+            if not hasattr(self.dataset.nodes, "get_nodes_in_selection_set"):
+                raise AttributeError(
+                    "self.dataset.nodes lacks 'get_nodes_in_selection_set'. "
+                    "Implement it or pass explicit 'node_ids'."
+                )
+            node_ids = self.dataset.nodes.get_nodes_in_selection_set(selection_set_id)
+
+        # preserve caller order, drop duplicates
+        node_ids = list(dict.fromkeys(node_ids))
+
+        # --- master table ------------------------------------------------- #
+        df_all = (
+            self.dataset.nodes_info["dataframe"]
+            if isinstance(self.dataset.nodes_info, dict)
+            else self.dataset.nodes_info
+        )
+
+        # --- verify existence -------------------------------------------- #
+        missing = set(node_ids) - set(df_all["node_id"])
+        if missing:
+            raise KeyError(f"Unknown node IDs: {sorted(missing)}")
+
+        # --- slice while keeping order ----------------------------------- #
+        sub = (
+            df_all.set_index("node_id")
+            .loc[node_ids]              # preserves specified order
+            .reset_index()
+        )
+
+        # --- return format ----------------------------------------------- #
+        if as_dict:
+            return {row.node_id: row._asdict() for row in sub.itertuples(index=False)}
+        return sub
+    # ─────────────────────────────────────────────────────────────────────────
+    @staticmethod
+    def _to_step_int(step_key: str | bytes, pattern: str = r"(\d+)$") -> int:
+        """
+        Convert an HDF5 dataset key to its integer STEP index.
+
+        Parameters
+        ----------
+        step_key : str | bytes
+            Raw dataset key from HDF5 (e.g. 'STEP_0', b'3', 'Step-12').
+        pattern : str, optional
+            Regex that captures the numeric portion; default grabs trailing digits.
+
+        Returns
+        -------
+        int
+            Numeric STEP value.
+        """
+        if isinstance(step_key, bytes):           # h5py may yield bytes
+            step_key = step_key.decode()
+
+        # Fast path: key is already numeric
+        if step_key.isdigit():
+            return int(step_key)
+
+        # Fallback: extract digits with regex
+        match = re.search(pattern, step_key)
+        if match:
+            return int(match.group(1))
+
+        raise ValueError(f"Un-recognisable STEP key: {step_key!r}")
+
+
+# The legacy ``ModelInfo`` alias lives on the ``STKO_to_python.model``
+# package surface (quiet) and at the deep path
+# ``STKO_to_python.model.model_info`` (DeprecationWarning via PEP 562
+# ``__getattr__`` shim). It is intentionally not declared on this
+# canonical module so that the library never trips its own warning.
+
+    
+    
+    
+    
+    
+    
+    
+    
+
